+++ instances/apps.py
@@ -22,6 +22,18 @@ def migrate_can_clone_instances(sender, **kwargs):
                         user.user_permissions.add(permission)
             break
 
+def apply_passwordless_console(sender, **kwargs):
+    '''
+    Apply new passwordless_console permission for all users
+    '''
+    from django.conf import settings
+    from django.contrib.auth.models import User, Permission
+
+    print('\033[92mApplying permission passwordless_console for all users\033[0m')
+    users = User.objects.all()
+    permission = Permission.objects.get(codename='passwordless_console')
+    for user in users:
+            user.user_permissions.add(permission)
 
 class InstancesConfig(AppConfig):
     name = 'instances'
@@ -29,3 +41,4 @@ class InstancesConfig(AppConfig):
 
     def ready(self):
         post_migrate.connect(migrate_can_clone_instances, sender=self)
+        post_migrate.connect(apply_passwordless_console, sender=self)
